chore: refactor kube deployment columns
Refactors the Kubernetes Deployment columns using /columns and a props.ts.

Part of #11035.
