+++ packages/renderer/src/lib/deployments/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import DeploymentActions from '../DeploymentActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<DeploymentActions deployment={object} />
