+++ packages/renderer/src/lib/deployments/columns/Conditions.spec.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2023-2024 Red Hat, Inc.
+ * Copyright (C) 2023-2025 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,8 +21,8 @@ import '@testing-library/jest-dom/vitest';
 import { render, screen } from '@testing-library/svelte';
 import { expect, test } from 'vitest';
 
-import DeploymentColumnConditions from './DeploymentColumnConditions.svelte';
-import type { DeploymentCondition, DeploymentUI } from './DeploymentUI';
+import type { DeploymentCondition, DeploymentUI } from '../DeploymentUI';
+import Conditions from './Conditions.svelte';
 
 function createDeploymentUI(conditions: DeploymentCondition[]): DeploymentUI {
   return {
@@ -41,7 +41,7 @@ test('Expect column styling available', async () => {
   const deployment = createDeploymentUI([
     { type: 'Available', message: 'Running fine', reason: 'MinimumReplicasAvailable' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Available');
   expect(text).toBeInTheDocument();
@@ -55,7 +55,7 @@ test('Expect column styling unavailable', async () => {
   const deployment = createDeploymentUI([
     { type: 'Available', message: 'Running fine', reason: 'MinimumReplicasUnavailable' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Unavailable');
   expect(text).toBeInTheDocument();
@@ -69,7 +69,7 @@ test('Expect column styling updated', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'ReplicaSetUpdated' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Updated');
   expect(text).toBeInTheDocument();
@@ -83,7 +83,7 @@ test('Expect column styling new replica set', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'NewReplicaSetCreated' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('New Replica Set');
   expect(text).toBeInTheDocument();
@@ -97,7 +97,7 @@ test('Expect column styling progressed', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'NewReplicaSetAvailable' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Progressed');
   expect(text).toBeInTheDocument();
@@ -111,7 +111,7 @@ test('Expect column styling scaled up', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'ReplicaSetScaledUp' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Scaled Up');
   expect(text).toBeInTheDocument();
@@ -125,7 +125,7 @@ test('Expect column styling scaled down', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'ReplicaSetScaledDown' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Scaled Down');
   expect(text).toBeInTheDocument();
@@ -139,7 +139,7 @@ test('Expect column styling deadline exceeded', async () => {
   const deployment = createDeploymentUI([
     { type: 'Progressing', message: 'Running fine', reason: 'ProgressDeadlineExceeded' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Deadline Exceeded');
   expect(text).toBeInTheDocument();
@@ -153,7 +153,7 @@ test('Expect column styling replica failure', async () => {
   const deployment = createDeploymentUI([
     { type: 'ReplicaFailure', message: 'Running fine', reason: 'ReplicaFailure' },
   ]);
-  render(DeploymentColumnConditions, { object: deployment });
+  render(Conditions, { object: deployment });
 
   const text = screen.getByText('Replica Failure');
   expect(text).toBeInTheDocument();
